Extract ClientOpts and AutoRetry to new file
