+++ client.go
@@ -11,7 +11,6 @@ import (
 	"os"
 	"os/signal"
 	"regexp"
-	"runtime"
 	"strings"
 	"sync"
 	"syscall"
@@ -20,61 +19,6 @@ import (
 	"github.com/AshokShau/gotdbot/internal/qrcode"
 	"github.com/AshokShau/gotdbot/internal/tdjson"
 )
-
-type AutoRetry struct {
-	ChatNotFound    bool
-	MessageNotFound bool
-}
-
-type ClientOpts struct {
-	LibraryPath             string
-	UseTestDC               bool
-	DatabaseDirectory       string
-	FilesDirectory          string
-	DatabaseEncryptionKey   string
-	UseFileDatabase         *bool
-	UseChatInfoDatabase     *bool
-	UseMessageDatabase      *bool
-	UseSecretChats          *bool
-	LoadMessagesBeforeReply bool
-	SystemLanguageCode      string
-	DeviceModel             string
-	SystemVersion           string
-	ApplicationVersion      string
-	TDLibOptions            map[string]interface{}
-	Logger                  *slog.Logger
-	QrMode                  bool
-	AuthorizationTimeout    time.Duration
-	LogVerbosityLevel       int32
-	LogStream               LogStream
-	AutoRetry               *AutoRetry
-
-	// Dispatcher is the dispatcher to use for this client.
-	// If nil, a new dispatcher will be created.
-	Dispatcher *Dispatcher
-}
-
-func DefaultClientConfig() *ClientOpts {
-	return &ClientOpts{
-		DatabaseDirectory:       "database",
-		FilesDirectory:          "",
-		DatabaseEncryptionKey:   "",
-		UseFileDatabase:         Bool(true),
-		UseChatInfoDatabase:     Bool(true),
-		UseMessageDatabase:      Bool(true),
-		UseSecretChats:          Bool(false),
-		LoadMessagesBeforeReply: false,
-		SystemLanguageCode:      "en",
-		DeviceModel:             "Gotdbot",
-		SystemVersion:           runtime.GOOS,
-		ApplicationVersion:      "Gotdbot " + Version,
-		Logger:                  slog.New(slog.NewTextHandler(os.Stdout, nil)),
-		QrMode:                  false,
-		AuthorizationTimeout:    60 * time.Second,
-		LogVerbosityLevel:       2,
-		AutoRetry:               &AutoRetry{},
-	}
-}
 
 type Client struct {
 	manager  *ClientManager
